requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.24.0.post1
httpx==0.25.2
google-cloud-bigquery==3.15.0
google-cloud-secret-manager==2.17.0
pydantic==1.10.13
python-dotenv==1.0.1
tenacity==8.2.3
python-dateutil==2.8.2
pytest==7.4.4
